Final Push (For Class)
Added final editions of the CRUD Page of the site
Known Error : Need to add a school object when adding a course. School cannot be Null leading to an error on a GET Request after.

Need to fix:
Final repository getAllCourse method filtering by Region, Prereq and School Id. Tries to use course ID as the sorting ID.